Start Changes to Deck for mathematical functionality
Deck.Original doesn't allow for mathematics - all cards are represented as strings. The first step in teaching the user is how to add mathematics to the existing card class - they will be editing the project as a follow up to a previous RPF project. 

This file will be the edited version that we will be showing to users and creating the working project with.